+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - the collected data is now flushed after each precondition 
 - moved test case specific parameter into separate file test_parameter.toml
-- sdc-ri version to 6.0.0-SNAPSHOT
+- sdc-ri version to 6.2.0-SNAPSHOT
 
 ### Fixed